+++ CHANGELOG.md
@@ -23,7 +23,11 @@ For more information on SemVer, please visit http://semver.org/.
 ## Version 1.7.3
 ### 2014-11-??
 
-* Fixed dynamic table css styles
+* Improved WPDK Preferences with `wpdk_preferences_reset_to_default_branch-{$branch}`.
+* Improved WPDK Preferences with `wpdk_preferences_update_branch-{$branch}`.
+* Added `wpdk_flush_cache_third_parties_plugins` action in order to flush third parties plugins.
+* Added `WPDKUIControlType::inputTypeWithClass()` helper static method in order to retrive the HTML control type string.
+* Fixed dynamic table css styles.
 
 
 ## Version 1.7.2